Extract filters component
